Make the intersection geometry logic an operation. Keep the transform for the moment. #136
