+++ osm2streets/src/operations/mod.rs
@@ -2,3 +2,4 @@
 
 mod collapse_intersection;
 mod collapse_short_road;
+mod update_geometry;
